Extract activity handler functions to their own module.
