+++ CataclysmWin.cbp
@@ -182,6 +182,8 @@
 		<Unit filename="src/action.h" />
 		<Unit filename="src/active_item_cache.cpp" />
 		<Unit filename="src/active_item_cache.h" />
+		<Unit filename="src/activity_handlers.cpp" />
+		<Unit filename="src/activity_handlers.h" />
 		<Unit filename="src/activity_item_handling.cpp" />
 		<Unit filename="src/addiction.cpp" />
 		<Unit filename="src/addiction.h" />
